+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 ## Unreleased
 
+- Engine features:
+  - Added character case assertion helpers `IsLower`, and `IsUpper`. [#555](https://github.com/microsoft/PSRule/issues/555)
+    - `IsLower` checks that all letters in a field value are lowercase.
+    - `IsUpper` checks that all letters in a field value are uppercase.
+
 ## v0.21.0-B2009006 (pre-release)
 
 What's changed since v0.20.0: